FastAPI/requirements.txt
fastapi
uvicorn
neo4j
paho-mqtt
pyjwt
python-dotenv
